src/requirements.txt
Flask==2.0.1	
Jinja2==3.0.1	
MarkupSafe==2.0.1	
Pillow==8.2.0		
Werkzeug==2.0.1	
certifi==2021.5.30	
chardet==4.0.0	
click==7.1.2		
docx==0.2.4	
idna==2.10	
itsdangerous==2.0.1	
lxml==4.6.3	
numpy==1.21.0	
pandas==1.2.5	
pip==21.1.2	
python-docx==0.8.11	
python-dotenv==0.17.1	
pytz==2021.1	
requests==2.25.1	
setuptools==57.0.0	
six==1.16.0	
urllib3==1.26.5	
wheel==0.36.2	